# docs/openapi/api-gateway.openapi.yaml
openapi: 3.0.3
x-last-modified: 2026-06-21
info:
  title: Rione API Gateway
  version: v0.3.0
  description: Public gateway contract for Rione service routes.
servers:
  - url: http://localhost:8080
    description: Local gateway
tags:
  - name: Users
    description: Proxied user service operations
  - name: Social
    description: Proxied social service operations under the /social prefix
paths:
  /users:
    post:
      tags:
        - Users
      operationId: gatewaySignUp
      summary: Create a user account
      description: Proxies POST /users to the user service.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
      responses:
        '201':
          description: User account created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogInResponse'
        '400':
          $ref: '#/components/responses/UserBadRequest'
  /users/login:
    post:
      tags:
        - Users
      operationId: gatewayLogIn
      summary: Log in a user
      description: Proxies POST /users/login to the user service.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogInRequest'
      responses:
        '200':
          description: User authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/UserBadRequest'
  /users/me:
    get:
      tags:
        - Users
      operationId: gatewayGetCurrentUser
      summary: Get the authenticated user profile
      description: Proxies GET /users/me to the user service.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/UserBadRequest'
  /users/me/profile:
    put:
      tags:
        - Users
      operationId: gatewayUpdateCurrentUserProfile
      summary: Update the authenticated user profile
      description: Proxies PUT /users/me/profile to the user service.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: User profile updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/UserBadRequest'
  /social/neighbor-requests:
    post:
      tags:
        - Social
      operationId: gatewaySendNeighborRequest
      summary: Send a neighbor request
      description: Proxies POST /social/neighbor-requests to POST /neighbor-requests in the social service.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendNeighborRequestRequest'
      responses:
        '201':
          description: Neighbor request created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NeighborRequestResponse'
        '400':
          $ref: '#/components/responses/SocialBadRequest'
  /social/neighbor-requests/{requestId}/acceptance:
    post:
      tags:
        - Social
      operationId: gatewayAcceptNeighborRequest
      summary: Accept a neighbor request
      description: Proxies the request to the social service after removing the /social prefix.
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Neighbor request accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NeighborRequestResponse'
        '400':
          $ref: '#/components/responses/SocialBadRequest'
  /social/neighbor-requests/{requestId}/rejection:
    post:
      tags:
        - Social
      operationId: gatewayRejectNeighborRequest
      summary: Reject a neighbor request
      description: Proxies the request to the social service after removing the /social prefix.
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Neighbor request rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NeighborRequestResponse'
        '400':
          $ref: '#/components/responses/SocialBadRequest'
  /social/me/neighbor-requests/sent:
    get:
      tags:
        - Social
      operationId: gatewayGetMySentNeighborRequests
      summary: List my sent neighbor requests
      description: Proxies to GET /me/neighbor-requests/sent in the social service.
      responses:
        '200':
          description: Neighbor requests sent by the user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NeighborRequestResponse'
        '400':
          $ref: '#/components/responses/SocialBadRequest'
  /social/me/neighbor-requests/received:
    get:
      tags:
        - Social
      operationId: gatewayGetMyReceivedNeighborRequests
      summary: List my received neighbor requests
      description: Proxies to GET /me/neighbor-requests/received in the social service.
      responses:
        '200':
          description: Neighbor requests received by the user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NeighborRequestResponse'
        '400':
          $ref: '#/components/responses/SocialBadRequest'
  /social/me/neighborships:
    get:
      tags:
        - Social
      operationId: gatewayGetMyNeighbors
      summary: List my neighbors
      description: Proxies to GET /me/neighborships in the social service.
      responses:
        '200':
          description: Neighbors for the authenticated user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NeighborResponse'
        '400':
          $ref: '#/components/responses/SocialBadRequest'
  /social/me/neighborships/{neighborId}:
    delete:
      tags:
        - Social
      operationId: gatewayDeleteMyNeighborship
      summary: Remove one of my neighborships
      description: Proxies to DELETE /me/neighborships/{neighborId} in the social service.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/NeighborId'
      responses:
        '204':
          description: Neighborship removed.
        '400':
          $ref: '#/components/responses/SocialBadRequest'
  /social/me/blocks:
    get:
      tags:
        - Social
      operationId: gatewayGetMyBlocks
      summary: List users I blocked
      description: Proxies to GET /me/blocks in the social service.
      responses:
        '200':
          description: Blocks created by the authenticated user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BlockResponse'
        '400':
          $ref: '#/components/responses/SocialBadRequest'
  /social/me/blocks/{blockedId}:
    put:
      tags:
        - Social
      operationId: gatewayPutMyBlock
      summary: Block a user idempotently
      description: Proxies to PUT /me/blocks/{blockedId} in the social service without a request body.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BlockedId'
      responses:
        '200':
          description: Existing block returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockResponse'
        '201':
          description: Block created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockResponse'
        '400':
          $ref: '#/components/responses/SocialBadRequest'
    delete:
      tags:
        - Social
      operationId: gatewayDeleteMyBlock
      summary: Unblock a user idempotently
      description: Proxies to DELETE /me/blocks/{blockedId} in the social service without a request body.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BlockedId'
      responses:
        '204':
          description: User is unblocked.
        '400':
          $ref: '#/components/responses/SocialBadRequest'
  /social/blocks:
    post:
      deprecated: true
      tags:
        - Social
      operationId: gatewayBlockUser
      summary: Block a user
      description: Proxies POST /social/blocks to POST /blocks in the social service.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlockUserRequest'
      responses:
        '201':
          description: Block created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockResponse'
        '400':
          $ref: '#/components/responses/SocialBadRequest'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    NeighborId:
      name: neighborId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    BlockedId:
      name: blockedId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    RequestId:
      name: requestId
      in: path
      required: true
      description: Neighbor request identifier.
      schema:
        type: integer
        format: int64
        minimum: 1
  responses:
    UserBadRequest:
      description: Invalid request or rejected user service rule.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UserErrorResponse'
    SocialBadRequest:
      description: Invalid request or rejected social service rule.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SocialErrorResponse'
  schemas:
    SignUpRequest:
      type: object
      required:
        - name
        - surname
        - username
        - mail
        - password
        - neighborhoodId
        - neighborhoodName
        - city
        - country
        - birthDate
      properties:
        name:
          type: string
          example: Mario
        surname:
          type: string
          example: Rossi
        username:
          type: string
          example: mario.rossi
        mail:
          type: string
          format: email
          example: mario.rossi@example.com
        password:
          type: string
          format: password
          writeOnly: true
        neighborhoodId:
          type: integer
          format: int64
          example: 10
        neighborhoodName:
          type: string
          example: San Lorenzo
        city:
          type: string
          example: Rome
        country:
          type: string
          example: Italy
        birthDate:
          type: string
          format: date-time
          example: "1990-05-10T00:00:00"
        bio:
          type: string
          nullable: true
    LogInRequest:
      type: object
      required:
        - mail
        - password
      properties:
        mail:
          type: string
          format: email
          example: mario.rossi@example.com
        password:
          type: string
          format: password
          writeOnly: true
    UpdateProfileRequest:
      type: object
      required:
        - name
        - surname
        - username
        - neighborhoodId
        - neighborhoodName
        - city
        - country
        - birthDate
      properties:
        name:
          type: string
          example: Mario
        surname:
          type: string
          example: Rossi
        username:
          type: string
          example: mario.rossi
        neighborhoodId:
          type: integer
          format: int64
          example: 10
        neighborhoodName:
          type: string
          example: San Lorenzo
        city:
          type: string
          example: Rome
        country:
          type: string
          example: Italy
        birthDate:
          type: string
          format: date-time
          example: "1990-05-10T00:00:00"
        bio:
          type: string
          nullable: true
    UserResponse:
      type: object
      required:
        - id
        - name
        - surname
        - username
        - mail
        - neighborhoodId
        - neighborhoodName
        - city
        - country
        - birthDate
        - admin
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        surname:
          type: string
        username:
          type: string
        mail:
          type: string
          format: email
        neighborhoodId:
          type: integer
          format: int64
        neighborhoodName:
          type: string
        city:
          type: string
        country:
          type: string
        birthDate:
          type: string
          format: date-time
        bio:
          type: string
          nullable: true
        admin:
          type: boolean
    LogInResponse:
      type: object
      required:
        - token
        - user
      properties:
        token:
          type: string
          description: JWT bearer token.
        user:
          $ref: '#/components/schemas/UserResponse'
    SendNeighborRequestRequest:
      type: object
      required:
        - receiverId
      properties:
        receiverId:
          type: integer
          format: int64
    BlockUserRequest:
      type: object
      required:
        - blockedId
      properties:
        blockedId:
          type: integer
          format: int64
    NeighborRequestResponse:
      type: object
      required:
        - id
        - senderId
        - receiverId
        - date
        - status
      properties:
        id:
          type: integer
          format: int64
        senderId:
          type: integer
          format: int64
        receiverId:
          type: integer
          format: int64
        date:
          type: string
          format: date-time
        status:
          type: string
          enum:
            - PENDING
            - ACCEPTED
            - REJECTED
    NeighborResponse:
      type: object
      required:
        - id
        - userId
        - neighborId
        - date
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        neighborId:
          type: integer
          format: int64
        date:
          type: string
          format: date-time
    BlockResponse:
      type: object
      required:
        - id
        - blockerId
        - blockedId
      properties:
        id:
          type: integer
          format: int64
        blockerId:
          type: integer
          format: int64
        blockedId:
          type: integer
          format: int64
    UserErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    SocialErrorResponse:
      type: object
      required:
        - message
        - timestamp
      properties:
        message:
          type: string
        timestamp:
          type: string
          format: date-time